\name{as.data.frame2.table}
\alias{as.data.frame2.table}
\title{Convert \code{table} a data.frame with repeated rows.}
\usage{
  as.data.frame2.table(x)
}
\arguments{
  \item{x}{an object of class \code{table}}
}
\value{
  A \code{data.frame}.
}
\description{
  Converts a \code{table} in a \code{data.frame} with rows
  consisting of the dimensions of the \code{table}, and the
  number rows as indicated by the \code{table}
}

