\name{intersection}
\alias{intersection}
\title{a multiple intersection function...}
\usage{intersection(x, y, ...)
intersection(x, y, ...)
}
\description{a multiple intersection function}
\details{...\code{intersection}: ...


}
\arguments{\item{x}{...}
\item{y}{...}
\item{...}{...}
}

\alias{intersection}

