\name{is.consistent.changeLevels}
\alias{is.consistent.changeLevels}
\title{Check levelChanges consistency...}
\usage{is.consistent.changeLevels(x, levelChanges,
    allowMissingCols)}

\description{
  Check levelChanges consistency
}

\details{
  Check that the levels provided in a levelChanges list
  match those that are already present.
}
\value{The data from \code{x[, levelChanges]}, with as prescribed by 
\code{levelChanges}.}
\arguments{
  \item{x}{a data frame}
  \item{levelChanges}{A list, of the form generated by 
\code{\link{generateFactorRenamer}}}
  \item{allowMissingCols}{A logical or a character vector. If it is a 
logical, it indicates whether to allow some columns of x 
not to be included in levelChanges. If TRUE, only a warning is issue if 
there are columns missing. If FALSE, an error is given.
If it is a character vector, it contains the names of any columns that 
are allowed not to be included in levelChanges.}
}
