\name{new_stack}
\alias{new_stack}
\title{Most-recently used stack...}
\usage{new_stack(size=1000)}

\description{
  Most-recently used stack
}

\details{
  Stack that is aware of which items have been used most
  recently.

  Based upon Hadley Wickham's memoise (MIT License)
  http://cran.r-project.org/web/packages/memoise/
}
\arguments{
  \item{size}{An integer. The stack size}
}
