\name{Rows2}
\alias{Rows2}
\title{Subset a nested list}
\usage{
  Rows2(x, which, size, use.names = F)
}
\arguments{
  \item{x}{a list}

  \item{which}{which element}

  \item{size}{the length of each sublist}

  \item{use.names}{logical, whether to use the names. It is
  faster not to use names.}
}
\value{
  a list
}
\description{
  Extract a particular element of a each sublist of a list.
}

