\name{strip.default}
\alias{strip.default}
\title{function title}
\usage{
  strip.default(which.given, which.panel, var.name,
  factor.levels, shingle.intervals = NULL, strip.names =
  c(FALSE, TRUE), strip.levels = c(TRUE, FALSE), sep = " :
  ", style = 1, horizontal = TRUE, bg =
  trellis.par.get("strip.background")$col[which.given], fg
  = trellis.par.get("strip.shingle")$col[which.given],
  par.strip.text = trellis.par.get("add.text"))
}
\arguments{
  \item{which.given}{...}

  \item{which.panel}{...}

  \item{var.name}{...}

  \item{factor.levels}{...}

  \item{shingle.intervals}{...}

  \item{strip.names}{...}

  \item{strip.levels}{...}

  \item{sep}{...}

  \item{style}{...}

  \item{horizontal}{...}

  \item{bg}{...}

  \item{fg}{...}

  \item{par.strip.text}{...}
}
\description{
  description
}

