\name{strip.default2}
\alias{strip.default2}
\title{strip.default2}
\usage{strip.default2(which.given, which.panel, var.name,
    factor.levels, shingle.intervals, strip.names=c(FALSE,
    TRUE), strip.levels=c(TRUE, FALSE), sep=" : ", style=1,
    horizontal=TRUE,
    bg=trellis.par.get("strip.background")$col[which.given],
    fg=trellis.par.get("strip.shingle")$col[which.given],
    par.strip.text=trellis.par.get("add.text"), dimx)}
