\name{useOuterStrips}
\alias{useOuterStrips}
\title{useOuterStrips...}
\usage{useOuterStrips(x, strip=strip.default, strip.left=strip.custom(horizontal = FALSE),
    strip.lines=1, strip.left.lines=strip.lines)
useOuterStrips(x, strip=strip.default, strip.left=strip.custom(horizontal = FALSE),
    strip.lines=1, strip.left.lines=strip.lines)
}
\description{useOuterStrips}
\details{...\code{useOuterStrips}: ...


}
\arguments{\item{x}{...}
\item{strip}{...}
\item{strip.left}{...}
\item{strip.lines}{...}
\item{strip.left.lines}{...}
}

\alias{useOuterStrips}

