\name{changeLevelsColName}
\alias{changeLevelsColName}
\title{Relevel a column using a levelChanges list...}
\usage{changeLevelsColName(colName, x, levelChanges, verbose=T)}

\description{
  Relevel a column using a levelChanges list
}

\details{
  Relevels a factor column of a data frame, as as
  prescribed by a level changes list, of the form generated
  by \code{\link{generateFactorRenamer}}.
}
\value{The data from \code{x[, levelChanges]}, with as prescribed by 
\code{levelChanges}.}
\arguments{
  \item{colName}{A character vector of length 1, indicating which column of 
\code{x} the string should be created for.}
  \item{x}{a data frame}
  \item{levelChanges}{A list, of the form generated by 
\code{\link{generateFactorRenamer}}}
  \item{verbose}{A logical, indicating whether to show progress}
}
