\name{Rows2}
\alias{Rows2}
\title{Subset a nested list...}
\usage{Rows2(x, which, size, use.names=F)}
\description{Subset a nested list}
\details{Extract a particular element of a each sublist of a list.}
\value{a list}
\arguments{\item{x}{a list}
\item{which}{which element}
\item{size}{the length of each sublist}
\item{use.names}{logical, whether to use the names. It is faster not to use names}}
