\name{useOuterStrips}
\alias{useOuterStrips}
\title{useOuterStrips}
\usage{
  useOuterStrips(x, strip = strip.default, strip.left =
  strip.custom(horizontal = FALSE), strip.lines = 1,
  strip.left.lines = strip.lines)
}
\arguments{
  \item{x}{...}

  \item{strip}{...}

  \item{strip.left}{...}

  \item{strip.lines}{...}

  \item{strip.left.lines}{...}
}
\description{
  ...
}

